// generated/cmd_pb.d.ts
// @generated by protoc-gen-es v2.2.5
// @generated from file cmd.proto (package wire, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { JsonObject, Message } from "@bufbuild/protobuf";
import type { Value } from "@bufbuild/protobuf/wkt";

/**
 * Describes the file cmd.proto.
 */
export declare const file_cmd: GenFile;

/**
 * @generated from message wire.Command
 */
export declare type Command = Message<"wire.Command"> & {
  /**
   * @generated from field: string cmd = 1;
   */
  cmd: string;

  /**
   * @generated from field: repeated string args = 2;
   */
  args: string[];
};

/**
 * Describes the message wire.Command.
 * Use `create(CommandSchema)` to create a new message.
 */
export declare const CommandSchema: GenMessage<Command>;

/**
 * @generated from message wire.Response
 */
export declare type Response = Message<"wire.Response"> & {
  /**
   * @generated from field: string err = 1;
   */
  err: string;

  /**
   * @generated from oneof wire.Response.value
   */
  value: {
    /**
     * @generated from field: bool v_nil = 2;
     */
    value: boolean;
    case: "vNil";
  } | {
    /**
     * @generated from field: int64 v_int = 3;
     */
    value: bigint;
    case: "vInt";
  } | {
    /**
     * @generated from field: string v_str = 4;
     */
    value: string;
    case: "vStr";
  } | {
    /**
     * @generated from field: double v_float = 5;
     */
    value: number;
    case: "vFloat";
  } | {
    /**
     * @generated from field: bytes v_bytes = 6;
     */
    value: Uint8Array;
    case: "vBytes";
  } | { case: undefined; value?: undefined };

  /**
   * @generated from field: google.protobuf.Struct attrs = 7;
   */
  attrs?: JsonObject;

  /**
   * @generated from field: repeated google.protobuf.Value v_list = 8;
   */
  vList: Value[];
};

/**
 * Describes the message wire.Response.
 * Use `create(ResponseSchema)` to create a new message.
 */
export declare const ResponseSchema: GenMessage<Response>;